requirements.txt
yt-dlp
mutagen
requests
google-auth-oauthlib
google-auth-httplib2